// extensions/matrix/src/matrix/sdk/types.ts
export interface MatrixRawEvent {
  event_id: string;
  sender: string;
  type: string;
  origin_server_ts: number;
  content: Record<string, unknown>;
  unsigned?: {
    age?: number;
    redacted_because?: unknown;
  };
  state_key?: string;
}